+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/auto2/AutoScanApp2.groovy
@@ -1,0 +1,26 @@
+package ru.vyarus.dropwizard.guice.support.auto2
+
+import io.dropwizard.Application
+import io.dropwizard.setup.Bootstrap
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.support.TestConfiguration
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 29.12.2022
+ */
+class AutoScanApp2 extends Application<TestConfiguration> {
+
+    @Override
+    void initialize(Bootstrap<TestConfiguration> bootstrap) {
+        bootstrap.addBundle(GuiceBundle.builder()
+                .enableAutoConfig()
+                .build()
+        );
+    }
+
+    @Override
+    void run(TestConfiguration configuration, Environment environment) throws Exception {
+    }
+}