+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - reusable application must be declared in base test class: all tests derived
     from this base class would use the same application instance
 * Add SBOM (json and xml with cyclonedx classifier)
+* Add .enableAutoConfig() no-args shortcut for enabling classpath scan in application package
 
 ### 5.6.1 (2022-07-02)
 * Update dropwizard to 2.1.1 (fixes java 8 issue by allowing afterburner usage)